+++ Changelog.md
@@ -2,8 +2,16 @@
 UserCandy Change Log
 https://www.usercandy.com
 ----------------------------------------------------------------------------------------
-UC v1.0.3
+UC v1.0.4
 ----------------------------------------------------------------------------------------
+### UC 1.0.4
+ - Added major security updates to the Dispenser
+ - Added assets folder ability to Dispenser Items
+ - Fixed error with deleting files on windows servers.
+ - Set Home, About, and Contact pages to copy to custom pages folder if not exist.
+ - Added fixes to Framework updater
+ - Changed when the custom/pages folder is created to fix error after install
+
 ### UC 1.0.3
  - Multiple Bug Fixes
  - Have Dispenser Framework updater fully working